Move abstract methods
