+++ _gulp/helpers.js
@@ -1,6 +1,11 @@
 const gutil = require('gulp-util');
 const path = require('path');
 const shell = require('shelljs');
+const recursive = require('recursive-readdir');
+const Promise = require('bluebird');
+const _ = require('lodash');
+
+const readFile = Promise.promisify(require('fs').readFile);
 
 const throwError = (plugin, error) => {
   throw new gutil.PluginError({
@@ -23,9 +28,51 @@ const touchFile = filePath => {
   }
 }
 
+const getFiles = (dir, ext) => {
+  const extName = ext || '.html';
+  return new Promise((resolve, reject) => {
+    recursive(dir, [], (err, files) => {
+      if (err) {
+        return reject(err);
+      }
+      resolve(_.filter(files, file => path.extname(file) === extName));
+    });
+  });
+};
+
+const readHtmlFile = filePath => {
+  return new Promise((resolve, reject) => {
+    readFile(filePath, "utf8").then(content => {
+      resolve({
+        path: filePath,
+        content: content,
+        links: [],
+        images:[],
+        anchors: [],
+        anchorLinks: [],
+        external: {
+          links: [],
+          images: [],
+          mailto: []
+        },
+        errors: [],
+        warnings: []
+      });
+    }).catch(reject);
+  });
+};
+
+const readHtmlFiles = paths => {
+  return Promise.all(_.map(paths, file => readHtmlFile(file)));
+};
+
 
 module.exports = {
   gulpErr: throwError,
   touch: touchFile,
-  isFile: isFile
+  isFile: isFile,
+  getFiles: getFiles,
+  readHtmlFile: readHtmlFile,
+  readHtmlFiles: readHtmlFiles,
+  readFile: readFile
 }
